@@ -1,65 +1,44 @@
 {
-  "name": "playground",
-  "version": "0.1.0",
-  "description": "",
-  "main": "index.js",
+  "name": "archive-client",
+  "type": "module",
+  "version": "0.0.1",
   "scripts": {
-    "dev": "webpack serve --port 3000",
-    "build": "NODE_ENV=production webpack",
-    "lint": "eslint -c .eslintrc.js src/",
-    "prettier": "prettier --config .prettierrc.js --write --check .",
-    "generate": "graphql-codegen --config codegen.yml"
-  },
-  "author": "",
-  "license": "ISC",
-  "devDependencies": {
-    "@svgr/webpack": "^6.2.1",
-    "@types/lodash": "^4.14.180",
-    "@types/react": "^17.0.39",
-    "@types/react-dom": "^17.0.11",
-    "@typescript-eslint/eslint-plugin": "^5.12.1",
-    "@typescript-eslint/parser": "^5.12.1",
-    "css-loader": "^6.6.0",
-    "eslint": "^8.10.0",
-    "eslint-config-prettier": "^8.4.0",
-    "eslint-plugin-prettier": "^4.0.0",
-    "eslint-plugin-react": "^7.29.2",
-    "eslint-plugin-react-hooks": "^4.3.0",
-    "file-loader": "^6.2.0",
-    "graphql-tag": "^2.12.6",
-    "html-webpack-plugin": "^5.5.0",
-    "mini-css-extract-plugin": "^2.5.3",
-    "postcss": "^8.4.12",
-    "postcss-loader": "^6.2.1",
-    "postcss-normalize": "^10.0.1",
-    "postcss-preset-env": "^7.4.3",
-    "prettier": "2.5.1",
-    "prettier-plugin-jsdoc": "^0.3.31",
-    "resolve-url-loader": "^5.0.0",
-    "rimraf": "^3.0.0",
-    "sass-loader": "^12.6.0",
-    "ts-loader": "^9.2.6",
-    "typescript": "^4.5.5",
-    "webpack": "^5.69.1",
-    "webpack-cli": "^4.9.2",
-    "webpack-dev-server": "^4.7.4",
-    "@graphql-codegen/typescript-operations": "2.3.7",
-    "@graphql-codegen/typescript": "2.4.10",
-    "@graphql-codegen/typescript-react-apollo": "3.2.13",
-    "@graphql-codegen/cli": "2.6.2"
+    "dev": "astro dev",
+    "start": "astro dev",
+    "build": "astro build",
+    "preview": "astro preview",
+    "astro": "astro",
+    "lint": "eslint -c .eslintrc.cjs src/",
+    "prettier": "prettier -c .prettierrc.cjs --write --check 'src/**/*.*'",
+    "generate": "graphql-codegen --config codegen.ts"
   },
   "dependencies": {
-    "@apollo/client": "^3.6.2",
-    "@graphql-codegen/cli": "^2.6.2",
-    "debounce": "^1.2.1",
-    "graphql": "^16.4.0",
-    "lodash": "^4.17.21",
-    "match-sorter": "^6.3.1",
-    "nanoid": "^3.3.4",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-router-dom": "^6.3.0",
-    "sass": "^1.49.9",
-    "use-debounce": "^7.0.1"
+    "@astrojs/node": "^6.0.3",
+    "@astrojs/svelte": "^4.0.2",
+    "@urql/svelte": "^4.0.4",
+    "astro": "^3.2.0",
+    "graphql": "^16.8.1",
+    "svelte": "^4.2.1",
+    "tint": "file:../../tint",
+    "urql": "^4.0.5"
+  },
+  "devDependencies": {
+    "@graphql-codegen/cli": "5.0.0",
+    "@graphql-codegen/client-preset": "^4.1.0",
+    "@graphql-codegen/typescript": "4.0.1",
+    "@graphql-codegen/typescript-operations": "^4.0.1",
+    "@graphql-codegen/typescript-resolvers": "^4.0.1",
+    "@graphql-codegen/typescript-urql": "^4.0.0",
+    "@trivago/prettier-plugin-sort-imports": "^4.2.0",
+    "@typescript-eslint/eslint-plugin": "^6.7.3",
+    "@typescript-eslint/parser": "^6.7.3",
+    "eslint": "^8.50.0",
+    "eslint-config-prettier": "^9.0.0",
+    "eslint-plugin-svelte": "^2.33.2",
+    "prettier": "^3.0.3",
+    "prettier-plugin-astro": "^0.12.0",
+    "prettier-plugin-jsdoc": "^1.0.2",
+    "prettier-plugin-svelte": "^3.0.3",
+    "vite-plugin-graphql-loader": "^2.0.0"
   }
 }
