@@ -1,67 +1,67 @@
 {
-  "name": "4x-archive",
+  "name": "archive-server",
   "version": "0.0.1",
+  "type": "module",
   "scripts": {
-    "dev": "dotenv -- nodemon src/index.ts",
+    "dev": "dotenv -- tsx watch src",
     "build": "rimraf dist && tsc",
     "lint": "eslint --ext .js,.ts -c .eslintrc.js src/",
     "prettier": "prettier --config .prettierrc.js --write --check .",
-    "knex": "dotenv -- knex",
+    "knex": "dotenv -- node --loader tsx ./node_modules/.bin/knex",
     "test": "dotenv -- ava"
   },
   "dependencies": {
-    "apollo-server-express": "^3.7.0",
-    "async-mutex": "^0.3.2",
-    "bcryptjs": "2.4.3",
-    "cors": "^2.8.5",
+    "@apollo/server": "^4.9.4",
+    "argon2": "^0.31.1",
+    "async-mutex": "^0.4.0",
     "cookie-parser": "^1.4.6",
-    "dataloader": "^2.1.0",
-    "express": "^4.18.1",
-    "file-type": "^17.1.1",
+    "cors": "^2.8.5",
+    "dataloader": "^2.2.2",
+    "express": "^4.18.2",
+    "file-type": "^18.5.0",
     "fluent-ffmpeg": "^2.1.2",
-    "fs-capacitor": "^7.0.1",
-    "fs-jetpack": "^4.3.1",
-    "graphql": "16.4.0",
-    "graphql-playground-middleware-express": "^1.7.23",
+    "fs-capacitor": "^8.0.0",
+    "fs-jetpack": "^5.1.0",
+    "graphql": "^16.8.1",
     "graphql-relay": "^0.10.0",
     "graphql-subscriptions": "^2.0.0",
-    "graphql-upload": "^13.0.0",
-    "graphql-ws": "^5.8.1",
-    "hashids": "^2.2.10",
-    "knex": "^2.0.0",
+    "graphql-upload": "^16.0.2",
+    "graphql-ws": "^5.14.1",
+    "hashids": "^2.3.0",
+    "knex": "^3.0.1",
     "morgan": "^1.10.0",
     "node-vibrant": "^3.1.6",
-    "objection": "^3.0.1",
+    "objection": "^3.1.2",
     "objection-unique": "^1.2.2",
-    "pg": "^8.7.3",
+    "pg": "^8.11.3",
     "pg-escape": "^0.2.0",
     "pg-ipc": "^1.0.5",
-    "sharp": "^0.30.4",
-    "string-strip-html": "^8.3.0",
-    "telegraf": "^4.8.1",
+    "sharp": "^0.32.6",
+    "string-strip-html": "^13.4.3",
+    "telegraf": "^4.14.0",
     "tmp": "^0.2.1"
   },
   "devDependencies": {
-    "@types/bcryptjs": "2.4.2",
-    "@types/body-parser": "^1.19.2",
-    "@types/cookie-parser": "^1.4.3",
-    "@types/express": "4.17.13",
-    "@types/graphql-upload": "^8.0.11",
-    "@types/morgan": "^1.9.3",
-    "@types/node": "17.0.31",
-    "@types/sharp": "^0.30.2",
-    "@types/tmp": "^0.2.3",
-    "@typescript-eslint/eslint-plugin": "^5.22.0",
-    "@typescript-eslint/parser": "^5.22.0",
-    "dotenv-cli": "5.1.0",
-    "eslint": "^8.15.0",
-    "eslint-config-prettier": "^8.5.0",
-    "eslint-plugin-jsdoc": "^39.2.9",
-    "nodemon": "2.0.16",
-    "prettier": "^2.6.2",
-    "prettier-plugin-jsdoc": "^0.3.38",
-    "ts-node": "10.7.0",
-    "tsconfig-paths": "^3.12.0",
-    "typescript": "4.6.4"
+    "@types/body-parser": "^1.19.3",
+    "@types/cookie-parser": "^1.4.4",
+    "@types/cors": "^2.8.14",
+    "@types/express": "4.17.18",
+    "@types/graphql-upload": "^16.0.2",
+    "@types/morgan": "^1.9.6",
+    "@types/node": "20.8.4",
+    "@types/sharp": "^0.31.1",
+    "@types/tmp": "^0.2.4",
+    "@typescript-eslint/eslint-plugin": "^6.7.5",
+    "@typescript-eslint/parser": "^6.7.5",
+    "dotenv-cli": "7.3.0",
+    "eslint": "^8.51.0",
+    "eslint-config-prettier": "^9.0.0",
+    "eslint-plugin-jsdoc": "^46.8.2",
+    "prettier": "^3.0.3",
+    "prettier-plugin-jsdoc": "^1.1.1",
+    "ts-node": "^10.9.1",
+    "tsconfig-paths": "^4.2.0",
+    "tsx": "^3.13.0",
+    "typescript": "5.2.2"
   }
 }
