@@ -1,19 +1,17 @@
-import express, { Request, Response } from 'express'
-import expressPlayground from 'graphql-playground-middleware-express'
-import { ApolloServer } from 'apollo-server-express'
+import express, { Request, Response, json } from 'express'
 import { Server as HttpServer } from 'http'
 import schema from './schema'
 import AuthCookieUtils from './AuthCookieUtils'
 import SessionActions from '@src/actions/SessionActions'
-import { makeExecutableSchema } from '@graphql-tools/schema'
 import { GraphQLError } from 'graphql'
 import Context from '@src/Context'
-import {
-  ApolloServerPluginDrainHttpServer,
-  ApolloServerPluginLandingPageDisabled,
-} from 'apollo-server-core'
 import { ValidationError } from 'objection'
 import { ValidationInputError } from '@src/errors'
+import { ApolloServer } from '@apollo/server'
+import { expressMiddleware } from '@apollo/server/express4';
+import { ApolloServerPluginDrainHttpServer } from '@apollo/server/plugin/drainHttpServer'
+// import { ApolloServerPluginLandingPageDisabled } from '@apollo/server/plugin/disabled'
+
 
 /**
  * This class is responsible for handling all GraphQL requests.
@@ -22,53 +20,42 @@ import { ValidationInputError } from '@src/errors'
  */
 export default class {
   async init(app: express.Application, server: HttpServer, options: any) {
-    // If server is in development mode, provide GraphQL Playground
-    if (process.env.NODE_ENV === 'development') {
-      app.use(
-        `${options.endpoint.replace(/^\/+/, '')}/playground`,
-        expressPlayground({
-          endpoint: options.endpoint,
-        }),
-      )
-    }
+    const apollo = new ApolloServer<Context>({
+      schema,
+      // formatError: (err) => {
+      //   if (err.originalError instanceof ValidationError) {
+      //     const newErr = new ValidationInputError(err.originalError)
+      //     return new GraphQLError(
+      //       newErr.message,
+      //       err.nodes,
+      //       err.source,
+      //       err.positions,
+      //       err.path,
+      //       newErr,
+      //       newErr.extensions,
+      //     )
+      //   }
 
-    const apollo = new ApolloServer({
-      schema: makeExecutableSchema({
-        typeDefs: schema,
-      }),
-      context: async ({ req, res }) =>
-        await this.createContext(req as any, res as any),
-      formatError: (err) => {
-        if (err.originalError instanceof ValidationError) {
-          const newErr = new ValidationInputError(err.originalError)
-          return new GraphQLError(
-            newErr.message,
-            err.nodes,
-            err.source,
-            err.positions,
-            err.path,
-            newErr,
-            newErr.extensions,
-          )
-        }
+      //   if (process.env.NODE_ENV !== 'development') {
+      //     delete err.extensions.exception
+      //   }
 
-        if (process.env.NODE_ENV !== 'development') {
-          delete err.extensions.exception
-        }
-
-        return err
-      },
+      //   return err
+      // },
       plugins: [
         ApolloServerPluginDrainHttpServer({ httpServer: server }),
-        ApolloServerPluginLandingPageDisabled(),
+        // ApolloServerPluginLandingPageDisabled(),
       ],
     })
     await apollo.start()
-    apollo.applyMiddleware({
-      app,
-      path: options.endpoint,
-      cors: options.corsOptions,
-    })
+
+    app.use(
+      options.endpoint,
+      json(),
+      expressMiddleware(apollo, {
+        context: async ({ req, res }) => await this.createContext(req as any, res as any),
+      }),
+    )
   }
 
   /**
