@@ -1,10 +1,11 @@
-import { ApolloError } from 'apollo-server-core'
+import { GraphQLError } from 'graphql'
 
-export default class AuthenticationError extends ApolloError {
+export default class AuthenticationError extends GraphQLError {
   constructor(msg?: string) {
-    super(
-      msg || 'You can only do this while logged-in.',
-      'AUTHENTICATION_ERROR',
-    )
+    super(msg || 'You can only do this while logged-in.', {
+      extensions: {
+        code: 'AUTHENTICATION_ERROR',
+      },
+    })
   }
 }
