@@ -1,7 +1,11 @@
-import { ApolloError } from 'apollo-server-core'
+import { GraphQLError } from 'graphql'
 
-export default class AuthorizationError extends ApolloError {
+export default class AuthorizationError extends GraphQLError {
   constructor(msg?: string) {
-    super(msg || 'You are not authorized to do this.', 'AUTHORIZATION_ERROR')
+    super(msg || 'You are not authorized to do this.', {
+      extensions: {
+        code: 'AUTHORIZATION_ERROR',
+      },
+    })
   }
 }
