@@ -1,7 +1,11 @@
-import { ApolloError } from 'apollo-server-core'
+import { GraphQLError } from 'graphql'
 
-export default class InputError extends ApolloError {
+export default class InputError extends GraphQLError {
   constructor(message: string) {
-    super(message, 'INPUT_ERROR')
+    super(message, {
+      extensions: {
+        code: 'INPUT_ERROR',
+      },
+    })
   }
 }
