@@ -95,5 +95,5 @@ export default class TaskModel extends BaseModel {
       },
     },
   }
-  static modelPaths = [__dirname]
+  static modelPaths = [new URL('.', import.meta.url).pathname]
 }
